feat: add

+++ groebner/composeApp/src/commonMain/kotlin/types/Monomial.kt
@@ -1,3 +1,3 @@
 package types
 
-data class Monomial(var coefficient: Long, val powers: HashMap<Char, UInt>)
+data class Monomial(var coefficient: Double, val powers: HashMap<Char, UInt>)
